improved the login and routes, to use the same login view for the authentication of both page and rest api
