+++ webcurator-ui/src/layout/MainLayoutView.vue
@@ -1,13 +1,15 @@
 <script setup lang="ts">
 import { RouterView } from 'vue-router';
-import LoginDialog from '@/views/login/LoginDialog.vue';
+import LoginView from '@/views/login/LoginView.vue';
 import NavBar from '@/components/NavBar.vue';
-import { useApiAuthStore } from '@/utils/rest.api';
-const authStore = useApiAuthStore();
+import { useAuthStore } from '@/utils/rest.api';
+const authStore = useAuthStore();
 </script>
 
 <template>
-  <LoginDialog v-if="authStore.isAuthenticating" id="login-dialog" />
+  <div v-if="authStore.isAuthenticating" id="login-dialog">
+    <LoginView />
+  </div>
   <div>
     <NavBar />
     <div class="main-container">
@@ -19,6 +21,9 @@ const authStore = useApiAuthStore();
 
 <style>
 #login-dialog {
+  width: 100vw;
+  height: 100vh;
+  background: var(--p-content-background);
   position: fixed;
   z-index: 9999;
 }
