--- webcurator-ui/src/views/login/LoginDialog.vue
@@ -1,9 +1,0 @@
-<script setup lang="ts">
-import LoginWidget from './LoginWidget.vue';
-</script>
-
-<template>
-  <div style="width: 100vw; height: 100vh; background: var(--p-content-background)">
-    <LoginWidget :redirectToLoginPage="false" />
-  </div>
-</template>
